Implement sessions in padwoman
In preparation for the next commit, which implements session cleanup on
logout, we need the ability to map the etherpad sessions to a padwoman
session and not just the user.
The reason for this is that not all sessions are logged out of etherpad
at once. So a logout on one pc does not log out another session of the
same user.