[1/4] Make public and lastEdit in padlist lazy loading
This commit is part of a series which goal is to minimize the loading
time of the padlist.

The main reason for slowness was that if the public status or lastEdit
timestamp where not cached in redis, padwoman would crawl them for the
whole pad group from the slow etherpad API before replying with the
page. That resulted unacceptable minute-long loading times.

This commit series will make the page return instantly, omitting the
data not in the redis cache. The remaining data will be fetched from the
etherpad API asynchronously and will be available in subsequent
requests. With JavaScript it is also possible to lazily load the data
without refreshing the page.

This commit removes the periodically scheduled updateTimestamps job. It
will not be necessary anymore.