[2/4] Make public and lastEdit in padlist lazy loading
This commit is part of a series which goal is to minimize the loading
time of the padlist.
See 93861328b4a73759207b82b73149ce51a245a831 for more explanation.

This commit reworks etherpad_cached_api.getPadlist so that it can return
very fast without talking to the etherpad API. If the necessary data
(public, lastEdit) is not in the redis cache, "Unavailable" is used
instead. In the padlist this is represented by a loading symbol.

If data needs to be fetched, a new thread is started that does this. A
mutex is used, so that only one thread fetches data.

If the padlist is sorted by a column with "Unavailable" data, they are
sorted as most recent (first in descending order). This is probably
true, because of the cache time heuristic. As a side effect of the
implementation, it is now possible to sort by the public status.