[3/4] Make public and lastEdit in padlist lazy loading
This commit is part of a series which goal is to minimize the loading
time of the padlist.
See 93861328b4a73759207b82b73149ce51a245a831 for more explanation.

This commit adds a microapi endpoint for getting the padlist. It is
synchronous and queries all missing data from the etherpad API. The
endpoint is used by the client via JavaScript to reload the page when
all data is fetched.