[4/4] Make public and lastEdit in padlist lazy loading
This commit is part of a series which goal is to minimize the loading
time of the padlist.
See 93861328b4a73759207b82b73149ce51a245a831 for more explanation.

This commit properly implements the client side of the lazy loading. The
data from the microapi reply is inserted into the padlist via
JavaScript. The page does not need to be reloaded.